backend/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
langchain
langchain-community
faiss-cpu
